# backend/internal/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Create Go App API
  version: 0.1.0
  description: Contract for health and documentation endpoints.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security: []
paths:
...
  /auth/login:
    post:
      operationId: postAuthLogin
      summary: Login and issue auth token
      description: Validates user credentials, returns JWT token in response body, and sets httpOnly auth_token cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: correct-password
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Sets auth_token httpOnly cookie for browser clients.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
              example:
                error: invalid credentials
        "422":
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorsResponse'
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
              example:
                error: database unavailable
  /auth/refresh:
    post:
      operationId: postAuthRefresh
      summary: Refresh auth token
      description: Accepts Bearer token or auth_token cookie, issues a new token, and resets the auth_token cookie.
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: Resets auth_token httpOnly cookie with refreshed token.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
              example:
                error: invalid token
  /auth/logout:
    post:
      operationId: postAuthLogout
      summary: Logout current browser session
      description: Clears auth_token cookie and always returns success.
      responses:
        "200":
          description: Logged out
          headers:
            Set-Cookie:
              description: Clears auth_token cookie by setting Max-Age to -1.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: logged out
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
              example:
                error: invalid request
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Endpoints accept JWT through Authorization header Bearer <token>.
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Endpoints also accept httpOnly auth_token cookie as an alternative to BearerAuth.
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: correct-password
    TokenResponse:
      type: object
      required:
        - token
        - expires_at
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_at:
          type: string
          format: date-time
          example: "2026-01-01T01:00:00Z"
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: logged out
    AuthErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid token
...
    ValidationErrorsResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: string
          example:
            - email is required
            - password is required
...